refactor: 추천 테스트 관련 파일 test 경로 아래로 응집
