+++ src/test/java/com/techfork/domain/recommendation/evaluation/RecommendationConfigComparisonTest.java
@@ -1,4 +1,4 @@
-package com.techfork.domain.recommendation;
+package com.techfork.domain.recommendation.evaluation;
 
 import com.techfork.domain.user.entity.User;
 import lombok.extern.slf4j.Slf4j;
