+++ src/test/java/com/techfork/domain/recommendation/setup/components/TestDataGenerator.java
@@ -3,9 +3,9 @@ package com.techfork.domain.recommendation.setup.components;
 import com.techfork.domain.activity.repository.ReadPostRepository;
 import com.techfork.domain.post.entity.Post;
 import com.techfork.domain.post.repository.PostRepository;
-import com.techfork.domain.recommendation_quality.ImprovedRecommendationTestCase;
-import com.techfork.domain.recommendation_quality.TrainTestSplit;
-import com.techfork.domain.recommendation_quality.UserCreationResult;
+import com.techfork.domain.recommendation.dto.ImprovedRecommendationTestCase;
+import com.techfork.domain.recommendation.dto.TrainTestSplit;
+import com.techfork.domain.recommendation.dto.UserCreationResult;
 import com.techfork.domain.user.document.UserProfileDocument;
 import com.techfork.domain.user.entity.User;
 import com.techfork.domain.user.enums.EInterestCategory;
